git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene5650@1596478 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/util/fst/ReverseBytesReader.java
@@ -39,7 +39,7 @@ final class ReverseBytesReader extends FST.BytesReader {
   }
 
   @Override
-  public void skipBytes(int count) {
+  public void skipBytes(long count) {
     pos -= count;
   }
